requirements.txt
amqp==5.1.1
celery==5.2.7
Django==4.0.6
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
djangorestframework==3.13.1
drf-yasg==1.20.0
python-decouple==3.6
redis==4.3.4
requests==2.28.1